cdrtools-3.00

+++ libschily/findbytes.c
@@ -1,45 +1,45 @@
-/* @(#)findbytes.c	1.2 03/06/15 Copyright 2000-2003 J. Schilling */
+/* @(#)findbytes.c	1.6 09/10/17 Copyright 2000-2009 J. Schilling */
 /*
  *	Find a byte with specific value in memory.
  *
- *	Copyright (c) 2000-2003 J. Schilling
+ *	Copyright (c) 2000-2009 J. Schilling
  *
  *	Based on a strlen() idea from Torbjorn Granlund (tege@sics.se) and
  *	Dan Sahlin (dan@sics.se) and the memchr() suggestion
  *	from Dick Karpinski (dick@cca.ucsf.edu).
  */
 /*
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2, or (at your option)
- * any later version.
+ * The contents of this file are subject to the terms of the
+ * Common Development and Distribution License, Version 1.0 only
+ * (the "License").  You may not use this file except in compliance
+ * with the License.
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * See the file CDDL.Schily.txt in this distribution for details.
  *
- * You should have received a copy of the GNU General Public License along with
- * this program; see the file COPYING.  If not, write to the Free Software
- * Foundation, 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
+ * When distributing Covered Code, include this CDDL HEADER in each
+ * file and include the License file CDDL.Schily.txt from this distribution.
  */
 
-#include <mconfig.h>
-#include <stdxlib.h>
-#include <utypes.h>
-#include <align.h>
-#include <standard.h>
-#include <strdefs.h>
-#include <schily.h>
+#include <schily/mconfig.h>
+#include <schily/stdlib.h>
+#include <schily/utypes.h>
+#include <schily/align.h>
+#include <schily/standard.h>
+#include <schily/string.h>
+#include <schily/types.h>
+#include <schily/schily.h>
 
+/*
+ * findbytes(p, cnt, val) is the same as memchr(p, val, cnt)
+ */
 #ifdef	PROTOTYPES
 EXPORT char *
-findbytes(const void *vp, int cnt, char val)
+findbytes(const void *vp, ssize_t cnt, char val)
 #else
 EXPORT char *
 findbytes(vp, cnt, val)
 		const	void	*vp;
-	register	int	cnt;
+	register	ssize_t	cnt;
 			char	val;
 #endif
 {
