+++ libschily/movebytes.c
@@ -1,40 +1,40 @@
-/* @(#)movebytes.c	1.13 03/06/15 Copyright 1985, 1989, 1995-2003 J. Schilling */
+/* @(#)movebytes.c	1.18 09/10/17 Copyright 1985, 1989, 1995-2009 J. Schilling */
 /*
  *	move data
  *
- *	Copyright (c) 1985, 1989, 1995-2003 J. Schilling
+ *	Copyright (c) 1985, 1989, 1995-2009 J. Schilling
  */
 /*
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2, or (at your option)
- * any later version.
+ * The contents of this file are subject to the terms of the
+ * Common Development and Distribution License, Version 1.0 only
+ * (the "License").  You may not use this file except in compliance
+ * with the License.
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * See the file CDDL.Schily.txt in this distribution for details.
  *
- * You should have received a copy of the GNU General Public License along with
- * this program; see the file COPYING.  If not, write to the Free Software
- * Foundation, 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
+ * When distributing Covered Code, include this CDDL HEADER in each
+ * file and include the License file CDDL.Schily.txt from this distribution.
  */
 
-#include <standard.h>
-#include <align.h>
-#include <schily.h>
+#include <schily/standard.h>
+#include <schily/align.h>
+#include <schily/types.h>
+#include <schily/schily.h>
 
 #define	DO8(a)	a; a; a; a; a; a; a; a;
 
+/*
+ * movebytes(from, to, cnt) is the same as memmove(to, from, cnt)
+ */
 EXPORT char *
 movebytes(fromv, tov, cnt)
 	const void	*fromv;
 	void		*tov;
-	int		cnt;
+	ssize_t		cnt;
 {
 	register const char	*from	= fromv;
 	register char		*to	= tov;
-	register int		n;
+	register ssize_t	n;
 
 	/*
 	 * If we change cnt to be unsigned, check for == instead of <=
@@ -44,14 +44,14 @@ movebytes(fromv, tov, cnt)
 
 	if (from >= to) {
 		/*
-		 * source is on higher adresses than destination:
+		 * source is on higher addresses than destination:
 		 *	move bytes forwards
 		 */
-		if (n >= (int)(8 * sizeof (long))) {
+		if (n >= (ssize_t)(8 * sizeof (long))) {
 			if (l2aligned(from, to)) {
 				register const long *froml = (const long *)from;
 				register long *tol = (long *)to;
-				register int rem = n % (8 * sizeof (long));
+				register ssize_t rem = n % (8 * sizeof (long));
 
 				n /= (8 * sizeof (long));
 				do {
@@ -84,17 +84,17 @@ movebytes(fromv, tov, cnt)
 		char *ep;
 
 		/*
-		 * source is on lower adresses than destination:
+		 * source is on lower addresses than destination:
 		 *	move bytes backwards
 		 */
 		to += n;
 		from += n;
 		ep = to;
-		if (n >= (int)(8 * sizeof (long))) {
+		if (n >= (ssize_t)(8 * sizeof (long))) {
 			if (l2aligned(from, to)) {
 				register const long *froml = (const long *)from;
 				register long *tol = (long *)to;
-				register int rem = n % (8 * sizeof (long));
+				register ssize_t rem = n % (8 * sizeof (long));
 
 				n /= (8 * sizeof (long));
 				do {
